refine: check consistency between preprocess and instrument
