+++ tool/shared/shared.go
@@ -25,6 +25,7 @@ import (
 	"strings"
 
 	"github.com/alibaba/opentelemetry-go-auto-instrumentation/tool/util"
+	"golang.org/x/mod/semver"
 )
 
 const GoBuildIgnoreComment = "//go:build ignore"
@@ -151,4 +152,59 @@ func GuaranteeInPreprocess() {
 
 func GuaranteeInInstrument() {
 	util.Assert(InToolexec, "not in instrument stage")
+}
+
+// splitVersionRange splits the version range into two parts, start and end.
+func splitVersionRange(vr string) (string, string) {
+	util.Assert(strings.Contains(vr, ","), "invalid version range format")
+	util.Assert(strings.Contains(vr, "["), "invalid version range format")
+	util.Assert(strings.Contains(vr, ")"), "invalid version range format")
+
+	start := vr[1:strings.Index(vr, ",")]
+	end := vr[strings.Index(vr, ",")+1 : len(vr)-1]
+	return "v" + start, "v" + end
+}
+
+var versionRegexp = regexp.MustCompile(`@v\d+\.\d+\.\d+(-.*?)?/`)
+
+func ExtractVersion(path string) string {
+	version := versionRegexp.FindString(path)
+	if version == "" {
+		return ""
+	}
+	// Extract version number from the string
+	return version[1 : len(version)-1]
+}
+
+// MatchVersion checks if the version string matches the version range in the
+// rule. The version range is in format [start, end), where start is inclusive
+// and end is exclusive. If the rule version string is empty, it always matches.
+func MatchVersion(version string, ruleVersion string) (bool, error) {
+	// Fast path, always match if the rule version is not specified
+	if ruleVersion == "" {
+		return true, nil
+	}
+	// Check if both rule version and package version are in sane
+	if !strings.Contains(version, "v") {
+		return false, fmt.Errorf("invalid version %v %v",
+			version, ruleVersion)
+	}
+	if !strings.Contains(ruleVersion, "[") ||
+		!strings.Contains(ruleVersion, ")") ||
+		!strings.Contains(ruleVersion, ",") ||
+		strings.Contains(ruleVersion, "v") {
+		return false, fmt.Errorf("invalid version format in rule %v",
+			ruleVersion)
+	}
+	// Remove extra whitespace from the rule version string
+	ruleVersion = strings.ReplaceAll(ruleVersion, " ", "")
+
+	// Compare the version with the rule version, the rule version is in the
+	// format [start, end), where start is inclusive and end is exclusive
+	ruleVersionStart, ruleVersionEnd := splitVersionRange(ruleVersion)
+	if semver.Compare(version, ruleVersionStart) >= 0 &&
+		semver.Compare(version, ruleVersionEnd) < 0 {
+		return true, nil
+	}
+	return false, nil
 }
